Test: 클래스 별로 테스트 분리
